pyproject.toml
[tool.poetry]
name = "llm-practice"
version = "0.1.0"
description = ""
authors = []

packages = [
    { include = "api", from = "bases" },
    { include = "health", from = "components" },
    { include = "otel", from = "components" },
]

package-mode = false

[tool.poetry.dependencies]
python = ">=3.12.*"
fastapi = "^0.110.1"
opentelemetry-api = "^1.24.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
opentelemetry-instrumentation-fastapi = "^0.45b0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pytest = "^8.1.1"
ruff = "^0.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
